More refactoring.
There should be no semantic changes here.
