+++ cryptol-repl-internal/REPL/Haskeline.hs
@@ -152,7 +152,7 @@ repl cryrc projectConfig replMode callStacks stopOnError begin =
        if crSuccess status then do
           begin
           case projectConfig of
-            Just config -> Project.loadProject config
+            Just config -> Project.loadProjectREPL config
             Nothing     -> crySession replMode stopOnError
        else return status
